Lec3: Java Variables

+++ src/lec02_java_basic/G_MustLearn.java
@@ -8,7 +8,7 @@ public class G_MustLearn { // starting of the class body
 	// Those variables can be either declared (no value assigned) or initialized (value assigned) 	
 	public String name; // variable declared
 	public String myName = "Tofael"; // variable initialized, variable type String, this represent any noun/description
-	public int myAge = 99; // variable type int, this represent number
+	public long myAge = 99; // variable type int, this represent number
 	public char mySex = 'M'; // variable type char, this represent character
 	public boolean usCitizen = true; // variable type boolean, this represent true or false
